// src/app/models/post.model.ts
export class Post {
    id: string = '';
    post: string = '';
    userId: string = '';
    userPhotoId: string = '';
    // imageId: string = '';
    postImageId: string = '';
    postImage: string = '';
    userName: string = '';
    postTimer: string = '';
    isAdmin: boolean = false;
    profession: string = '';
    isActive: boolean = false;
    isMyPost: boolean = false;
    isPostImage: boolean = false;
    userIcon: any;
}